+++ src/public/rest-api/class-projects.php
@@ -9,14 +9,11 @@ namespace PTC_Completionist\REST_API;
 
 defined( 'ABSPATH' ) || die();
 
-use const PTC_Completionist\PLUGIN_PATH;
 use const PTC_Completionist\REST_API_NAMESPACE_V1;
 
 use PTC_Completionist\Asana_Interface;
-use PTC_Completionist\Options;
 use PTC_Completionist\HTML_Builder;
 use PTC_Completionist\Request_Token;
-use PTC_Completionist\Util;
 
 /**
  * Class to register and handle custom REST API endpoints for Asana projects.